backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
redis>=5.0.1
python-dotenv>=1.0.1
pydantic-settings>=2.1.0
ccxt>=4.2.19
pandas>=2.2.0
httpx>=0.26.0
python-multipart>=0.0.7
websockets>=12.0
apscheduler>=3.10.4
feedparser>=6.0.11
web3>=6.15.0
eth-account>=0.11.0
pyjwt>=2.8.0
